move Elixir to own dir
