+++ packages/quicktype-core/src/language/Elixir/constants.ts
@@ -1,0 +1,101 @@
+export const forbiddenModuleNames = [
+    "Access",
+    "Agent",
+    "Any",
+    "Application",
+    "ArgumentError",
+    "ArithmeticError",
+    "Atom",
+    "BadArityError",
+    "BadBooleanError",
+    "BadFunctionError",
+    "BadMapError",
+    "BadStructError",
+    "Base",
+    "Behaviour",
+    "Bitwise",
+    "Calendar",
+    "CaseClauseError",
+    "Code",
+    "Collectable",
+    "CondClauseError",
+    "Config",
+    "Date",
+    "DateTime",
+    "Dict",
+    "DynamicSupervisor",
+    "Enum",
+    "ErlangError",
+    "Exception",
+    "File",
+    "Float",
+    "Function",
+    "FunctionClauseError",
+    "GenEvent",
+    "GenServer",
+    "HashDict",
+    "HashSet",
+    "IO",
+    "Inspect",
+    "Integer",
+    "Kernel",
+    "KeyError",
+    "Keyword",
+    "List",
+    "Macro",
+    "Map",
+    "MapSet",
+    "MatchError",
+    "Module",
+    "Node",
+    "OptionParser",
+    "Path",
+    "Port",
+    "Process",
+    "Protocol",
+    "Range",
+    "Record",
+    "Regex",
+    "Registry",
+    "RuntimeError",
+    "Set",
+    "Stream",
+    "String",
+    "StringIO",
+    "Supervisor",
+    "SyntaxError",
+    "System",
+    "SystemLimitError",
+    "Task",
+    "Time",
+    "TokenMissingError",
+    "Tuple",
+    "URI",
+    "UndefinedFunctionError",
+    "UnicodeConversionError",
+    "Version",
+    "WithClauseError"
+] as const;
+
+export const reservedWords = [
+    "def",
+    "defmodule",
+    "use",
+    "import",
+    "alias",
+    "true",
+    "false",
+    "nil",
+    "when",
+    "and",
+    "or",
+    "not",
+    "in",
+    "fn",
+    "do",
+    "end",
+    "catch",
+    "rescue",
+    "after",
+    "else"
+] as const;
