+++ packages/quicktype-core/src/language/Elixir/index.ts
@@ -1,0 +1,2 @@
+export { ElixirTargetLanguage, elixirOptions } from "./language";
+export { ElixirRenderer } from "./ElixirRenderer";
