+++ packages/quicktype-core/src/language/Elixir/utils.ts
@@ -1,0 +1,73 @@
+import unicode from "unicode-properties";
+
+import {
+    legalizeCharacters,
+    splitIntoWords,
+    combineWords,
+    firstUpperWordStyle,
+    allUpperWordStyle,
+    allLowerWordStyle,
+    utf32ConcatMap,
+    isPrintable,
+    escapeNonPrintableMapper,
+    intToHex,
+    isLetterOrUnderscore
+} from "../../support/Strings";
+
+function unicodeEscape(codePoint: number): string {
+    return `\\u{${intToHex(codePoint, 0)}}`;
+}
+
+export function capitalizeFirstLetter(str: string) {
+    return str.charAt(0).toUpperCase() + str.slice(1);
+}
+
+export const stringEscape = utf32ConcatMap(escapeNonPrintableMapper(isPrintable, unicodeEscape));
+
+export function escapeDoubleQuotes(str: string) {
+    return str.replace(/"/g, '\\"');
+}
+
+export function escapeNewLines(str: string) {
+    return str.replace(/\n/g, "\\n");
+}
+
+const isStartCharacter = isLetterOrUnderscore;
+
+function isPartCharacter(utf16Unit: number): boolean {
+    const category: string = unicode.getCategory(utf16Unit);
+    return ["Nd", "Pc", "Mn", "Mc"].indexOf(category) >= 0 || isStartCharacter(utf16Unit);
+}
+
+const legalizeName = legalizeCharacters(isPartCharacter);
+
+export function simpleNameStyle(original: string, uppercase: boolean): string {
+    if (/^[0-9]+$/.test(original)) {
+        original = `${original}N`;
+    }
+    const words = splitIntoWords(original);
+    return combineWords(
+        words,
+        legalizeName,
+        uppercase ? firstUpperWordStyle : allLowerWordStyle,
+        uppercase ? firstUpperWordStyle : allLowerWordStyle,
+        allUpperWordStyle,
+        allUpperWordStyle,
+        "",
+        isStartCharacter
+    );
+}
+
+export function memberNameStyle(original: string): string {
+    const words = splitIntoWords(original);
+    return combineWords(
+        words,
+        legalizeName,
+        allLowerWordStyle,
+        allLowerWordStyle,
+        allLowerWordStyle,
+        allLowerWordStyle,
+        "_",
+        isStartCharacter
+    );
+}
